chore(popover): split out Align and Justify
