+++ packages/Popover/src/index.ts
@@ -1,4 +1,7 @@
 import Popover from './Popover';
-export { Align, Justify, PopoverProps } from './Popover';
+import Align from './Align';
+import Justify from './Justify';
+export { PopoverProps } from './Popover';
 
 export default Popover;
+export { Align, Justify };
